+++ tests/Main.hs
@@ -8,6 +8,11 @@ import Data.Function ((&))
 import qualified Database.SQLite.Simple as SQL
 import Database.SQLite.Simple.QQ.Interpolated
 
+data Person = Person {name :: String, age :: Integer}
+
+instance SQL.ToRow Person where
+  toRow p = SQL.toRow (name p, age p)
+
 table :: String
 table = "people"
 
@@ -16,13 +21,18 @@ main = bracket (SQL.open ":memory:") SQL.close $ \conn -> do
   print [isql|CREATE TABLE !{table} (name TEXT, age INTEGER)|]
   conn & [iexecute|CREATE TABLE !{table} (name TEXT, age INTEGER)|]
 
-  print [isql|INSERT INTO !{table} VALUES ("clive", 40)|]
-  conn & [iexecute|INSERT INTO !{table} VALUES ("clive", 40)|]
+  let name = "clive"
+  print [isql|INSERT INTO !{table} VALUES ({name}, 40)|]
+  conn & [iexecute|INSERT INTO !{table} VALUES ({name}, 40)|]
 
-  ageSum <- conn & [ifold|SELECT age FROM !{table}|] 0 (\acc (SQL.Only x) -> pure (acc + x))
-  print (ageSum :: Int)
+  let clara = Person {name = "clara", age = 25}
+  print [isql|INSERT INTO !{table} VALUES @{clara}|]
+  conn & [iexecute|INSERT INTO !{table} VALUES @{clara}|]
 
-  let limit = 1 :: Int
-  print [isql|SELECT age FROM !{table} WHERE name = ${map toLower "CLIVE"} LIMIT ${limit}|]
-  ages <- conn & [iquery|SELECT age FROM !{table} WHERE name = ${map toLower "CLIVE"} LIMIT ${limit}|]
-  print (ages :: [SQL.Only Int])
+  ageHaskellSum <- conn & [ifold|SELECT age FROM !{table}|] 0 (\acc (SQL.Only x) -> pure (acc + x))
+  print (ageHaskellSum :: Int)
+
+  let minAge = 1 :: Int
+  print [isql|SELECT AVG(age) FROM !{table} WHERE age >= {minAge}|]
+  [ageAvg] <- conn & [iquery|SELECT AVG(age) FROM !{table} WHERE age >= {minAge}|]
+  print (ageAvg :: SQL.Only Double)